src/main/java/com/example/book/Book.java
package com.example.book;

import java.util.Date;
public class Book {
    private String title;
    private String author;
    private Date publicationDate;

    public Book(String title, String author, Date publicationDate) {
        this.title = title;
        this.author = author;
        this.publicationDate = publicationDate;
    }

    public String getTitle() {
        return title;
    }

    public String getAuthor() {
        return author;
    }

    public Date getPublicationDate() {
        return publicationDate;
    }
}
